added known walls

+++ src/main/java/ch/fhnw/FXView.java
@@ -510,7 +510,7 @@ class FXView {
                     map.getGrid().forEach(cellArrayList -> cellArrayList.forEach(cell -> {
                         if (cell.isStart()) {
                             cell.setStart(false);
-                            Pane currentStart = (Pane) getNodeFromGridPane(centerPane, cell.getIndex().i, cell.getIndex().j);
+                            Pane currentStart = (Pane) getNodeFromGridPane(centerPane, (int)cell.getIndex().i, (int)cell.getIndex().j);
                             if (currentStart != null) {
                                 currentStart.getStyleClass().removeAll("start");
                             }
@@ -544,7 +544,7 @@ class FXView {
         map.getPath().forEach(cell -> {
             if (cell.isVisited() && !cell.isWall() && !cell.isStart() && !cell.isCheckPoint()) {
 
-                Pane pane = (Pane) getNodeFromGridPane(centerPane, cell.getIndex().i, cell.getIndex().j);
+                Pane pane = (Pane) getNodeFromGridPane(centerPane, (int)cell.getIndex().i, (int)cell.getIndex().j);
                 if (pane != null) {
                     pane.getStyleClass().removeAll("visited");
                     pane.getStyleClass().add("path");
@@ -569,11 +569,11 @@ class FXView {
 
     void updateMap() {
         map.getGrid().forEach(cells -> cells.forEach(cell -> {
-            Pane currentPane = (Pane) getNodeFromGridPane(centerPane, cell.getIndex().i, cell.getIndex().j);
+            Pane currentPane = (Pane) getNodeFromGridPane(centerPane, (int)cell.getIndex().i, (int)cell.getIndex().j);
             if (currentPane != null) {
                 if (currentPane.getChildren().size() > 0) {
                     Label label = (Label) currentPane.getChildren().get(0);
-                    label.setText(String.valueOf(Math.round(map.getGrid().get(cell.getIndex().i).get(cell.getIndex().j).getF_score() * 100.0) / 100.0));
+                    label.setText(String.valueOf(Math.round(map.getGrid().get((int)cell.getIndex().i).get((int)cell.getIndex().j).getF_score() * 100.0) / 100.0));
                     if (cell.isWall()) {
                         label.setText(String.valueOf('\u221e'));
                         label.getStyleClass().add("white");
@@ -592,7 +592,7 @@ class FXView {
 
     void reset() {
         map.getGrid().forEach(cells -> cells.forEach(cell -> {
-            Pane currentPane = (Pane) getNodeFromGridPane(centerPane, cell.getIndex().i, cell.getIndex().j);
+            Pane currentPane = (Pane) getNodeFromGridPane(centerPane, (int)cell.getIndex().i, (int)cell.getIndex().j);
             if (currentPane != null) {
                 if (currentPane.getChildren().size() > 0) {
                     Label label = (Label) currentPane.getChildren().get(0);
